requirements.txt
aldryn-apphooks-config==0.3.3
aldryn-boilerplates==0.7.5
aldryn-bootstrap3==1.2.2
aldryn-categories==1.0.4
aldryn-common==1.0.4
aldryn-forms==3.0.3
aldryn-locations==1.0.1
aldryn-mailchimp==0.3.0
aldryn-newsblog==1.3.2
aldryn-people==1.2.2
aldryn-redirects==1.0.0
aldryn-reversion==1.1.0
aldryn-search==0.3.0
aldryn-translation-tools==0.2.1
aldryn-video==1.0.0
appdirs==1.4.3
asn1crypto==0.22.0
awesome-slugify==1.6.5
Babel==2.4.0
backport-collections==0.1
beautifulsoup4==4.5.3
braintree==3.37.2
cffi==1.10.0
cmsplugin-contact-plus==1.3.7
cmsplugin-filer==1.1.3
cryptography==2.0
cssselect==1.0.3
cssutils==1.0.2
defusedxml==0.5.0
diff-match-patch==20121119
Django==1.10.8
django-absolute==0.3
django-admin-sortable==2.0.21
django-allauth==0.31.0
django-appconf==1.0.2
django-appdata==0.1.6
django-autoslug==1.9.3
django-bootstrap-form==3.2.1
django-bootstrap3==8.2.3
django-carton==1.2.1
django-classy-tags==0.8.0
django-cleanup==0.4.2
django-cms==3.4.6
django-colorfield==0.1.15
django-countries==4.5
django-durationfield==0.5.2
django-emailit==0.2.3
django-filer==1.2.7
django-formtools==2.0
django-geoip2-extras==0.1.2
django-haystack==2.6.0
django-helpdesk==0.1.18
django-import-export==0.5.1
django-iprestrict==1.3.1
django-markdown-deux==1.0.5
django-meta==1.3.2
django-meta-mixin==0.3.0
django-mptt==0.8.7
django-parler==1.7
django-payments==0.11.0.3
django-phonenumber-field==1.3.0
django-polymorphic==1.0.2
django-ranged-response==0.2.0
django-recaptcha==1.4.0
django-redis==4.8.0
django-reversion==1.10.2
django-sekizai==0.10.0
Django-Select2==4.3.2
django-sendfile==0.3.11
django-simple-captcha==0.5.6
django-simple-math-captcha==1.0.7
django-sizefield==0.9.1
django-sortedm2m==1.3.3
django-spurl==0.6.4
django-standard-form==1.1.1
django-tablib==3.2
django-taggit==0.22.1
django-taggit-autosuggest==0.3.0
django-taggit-templatetags==0.2.5
django-templatetag-handlebars==1.3.1
django-templatetag-sugar==1.0
django-treebeard==4.1.2
django-twitter-tag==1.2.1
djangocms-admin-style==1.2.7
djangocms-apphook-setup==0.1.2
djangocms-attributes-field==0.3.0
djangocms-blog==0.8.10
djangocms-column==1.7.0
djangocms-googlemap==1.1.0
djangocms-link==2.1.2
djangocms-snippet==1.9.2
djangocms-style==2.0.2
djangocms-text-ckeditor==3.5.0
djangocms-video==2.0.3
easy-thumbnails==2.3
email-reply-parser==0.5.9
et-xmlfile==1.0.1
GeoIP==1.3.2
geoip2==2.4.0
hiredis==0.2.0
html5lib==0.9999999
idna==2.5
jdcal==1.3
jsonfield==1.0.3
lxml==3.7.3
mailchimp3==2.0.15
markdown2==2.3.3
maxminddb==1.2.3
micawber==0.3.3
mysqlclient==1.3.10
oauthlib==2.0.2
odfpy==1.3.5
olefile==0.44
openpyxl==2.4.8
packaging==16.8
phonenumbers==8.10.1
Pillow==4.2.1
pkg-resources==0.0.0
premailer==3.2.0
pycountry==17.5.14
pycparser==2.17
PyJWT==1.4.2
pyparsing==2.1.10
pyrate==0.5a6
python-dateutil==2.6.0
python-slugify==1.2.0
python3-openid==3.1.0
pytz==2018.4
PyYAML==3.12
redis==2.10.5
regex==2016.11.21
requests==2.14.2
requests-oauthlib==0.8.0
simplejson==3.10.0
six==1.10.0
sparkpost==1.3.2
stripe==1.51.0
suds-jurko==0.6
tablib==0.11.5
twitter==1.17.1
unicodecsv==0.14.1
Unidecode==0.4.20
URLObject==2.4.2
uwsgidecorators==1.1.0
xlrd==1.0.0
xlwt==1.2.0
xmltodict==0.10.2
YURL==0.13
